core/rawdb: change freezerTableMeta.flushOffset to int64
The flushOffset is a file position, which is usually tracked as int64 in Go.
It's better to keep it as the same type to avoid conversion.

Overall, this change removes some conversions, and introduces some.
The new conversions are in cases where flushOffset is compared
with numEntries*indexEntrySize computations.

+++ core/rawdb/freezer_meta_test.go
@@ -79,7 +79,7 @@ func TestUpgradeMetadata(t *testing.T) {
 	if meta.virtualTail != uint64(100) {
 		t.Fatal("Unexpected virtual tail field")
 	}
-	if meta.flushOffset != uint64(0) {
+	if meta.flushOffset != 0 {
 		t.Fatal("Unexpected flush offset field")
 	}
 
@@ -95,7 +95,7 @@ func TestUpgradeMetadata(t *testing.T) {
 	if meta.virtualTail != uint64(100) {
 		t.Fatal("Unexpected virtual tail field")
 	}
-	if meta.flushOffset != uint64(100) {
+	if meta.flushOffset != 100 {
 		t.Fatal("Unexpected flush offset field")
 	}
 }
